smt-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.4.6
pymongo==4.7.2
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.3
pydantic[email]==2.7.1
python-multipart==0.0.9
reportlab==4.1.0
